==> 2.Drugbank.ipynb <==
pip install xmltodict

import xmltodict, json
import pandas as pd, sqlite3 
from rdkit import Chem
from rdkit.Chem import AllChem, PandasTools 
from rdkit.Chem.Draw import IPythonConsole  

### Format conversion

* visit https://www.drugbank.ca/releases/latest
* download drugbank_all_full_database.xml.zip
* unzip and get full database.xml
* rename to drugbank.5.1.7.xml

with open("/scratch/xg590/source/drugbank_all_full_database.5.1.8.xml", 'r') as xml_file:
    data_dict = xmltodict.parse(xml_file.read()) 
# executed in 2m 33s, finished 04:46:32 2021-05-21

### Put into Database

con = sqlite3.connect('/scratch/xg590/scratch/drugbank.5.1.8.db') 
cur = con.cursor() 
cur.execute('CREATE TABLE IF NOT EXISTS drugbank(drugbank_id TEXT, InChIKey TEXT, groups TEXT)') 

for one_drug_entry in data_dict["drugbank"]['drug']:
    if 'calculated-properties' in one_drug_entry.keys(): 
        if one_drug_entry['calculated-properties'] == None: continue
        #try:
        if '#text' in one_drug_entry['drugbank-id']:
            DB_ID = one_drug_entry['drugbank-id']['#text']
        else:
            DB_ID = one_drug_entry['drugbank-id'][0]['#text']
        if 'kind' in one_drug_entry['calculated-properties']['property']:
            one_property = one_drug_entry['calculated-properties']['property'] 
        else:
            for one_property in one_drug_entry['calculated-properties']['property']:
                if one_property['kind'] == 'SMILES': 
                    break 
        groups = one_drug_entry['groups']['group']
        if isinstance(groups, list): 
            groups = '|'.join(groups)
        SMILES = one_property['value']
        mol = Chem.MolFromSmiles(SMILES)
        if mol: 
            cur.execute('INSERT INTO drugbank (drugbank_id, InChIKey, groups) VALUES(?, ?, ?);',
                        (DB_ID, Chem.MolToInchiKey(mol), groups))  

cur.execute('commit;')
cur.close()

### Compare

con = sqlite3.connect('/scratch/xg590/scratch/drugbank.5.1.8.db') 

pd.read_pickle('data/df_latest.pkl').drop_duplicates('InChIKey')[['InChIKey']].to_sql('covdb',con,if_exists="replace",index=False)

df = pd.read_sql('''SELECT covdb.InChIKey, drugbank.drugbank_id, groups 
               FROM drugbank JOIN covdb   
               ON drugbank.InChIKey=covdb.InChIKey;''', con) 

df.to_csv('data/drugbank.csv',index=False) 
with open('data/drugbank.csv','a') as fw:
    fw.write('RICZEKWVNZFTNZ-LFGITCQGSA-N,DB14760,\n')

pd.read_sql("select * from drugbank where drugbank_id='DB14760'", con)
