==> 3.SIFTS.ipynb <==
### Get all xml.gz from SIFTS

ftp -i ftp.ebi.ac.uk  
anonymous
cd pub/databases/msd/sifts/xml
mget *.xml.gz

# pip install pandas 
import re, os, zlib, time, json, pickle, sqlite3, urllib, urllib.request, subprocess, datetime, multiprocessing, pandas as pd
con = sqlite3.connect('/scratch/xg590/scratch/sifts.20220302.db')  
cur = con.cursor()

import urllib, zlib
from bs4 import BeautifulSoup
def sifts_get_mapping_by_pdbid(p): 
    with open(p, 'rb') as fr: 
        xml  = zlib.decompressobj(32+zlib.MAX_WBITS).decompress(fr.read()) 
        soup = BeautifulSoup(xml, "lxml")
        s    = ''
        for residue in soup.find_all('residue'): 
            UniProt = residue.find(dbsource="UniProt")
            PDB = residue.find(dbsource="PDB") 
            try:
                if UniProt:
                    s += '{},{},{},{},{},{},{}\n'.format(PDB['dbaccessionid'], PDB['dbchainid'], PDB['dbresnum'], PDB['dbresname'], UniProt['dbaccessionid'], UniProt['dbresnum'], UniProt['dbresname'])
                else:
                    s += '{},{},{},{},null,null,null\n'.format(PDB['dbaccessionid'], PDB['dbchainid'], PDB['dbresnum'], PDB['dbresname']) 
            except Exception as e:
                print(pdbid, residue)
                s += json.dumps(residue.attrs)
                s += '\n'
        return s

#sifts_get_mapping_by_pdbid('data/sifts/102l.xml.gz')[:60]

timestamp = datetime.datetime.fromtimestamp(time.time()).strftime("%Y%m%d_%H%M%S")
sifts_csv = f'data/sifts.mining.{timestamp}.csv'

all_sifts_xml_gz = [p.path for p in os.scandir('data/sifts') if p.is_file()] 

num_worker = multiprocessing.cpu_count()
with multiprocessing.Pool(num_worker) as p, open('sifts.mining.progress', 'w') as fw:  
    num_jobs = len(all_sifts_xml_gz)
    jobs = [p.apply_async(sifts_get_mapping_by_pdbid, (i,)) for i in all_sifts_xml_gz]
    while True:
        time.sleep(60) 
        num_ready = len([job.ready() for job in jobs if job.ready()])
        print(f'Finish rate: {num_ready} / {num_jobs} @ {datetime.datetime.now().isoformat()}', file=fw)
        fw.flush()
        if num_ready == num_jobs:
            result = [job.get() for job in jobs]
            with open(sifts_csv, 'w') as _fw: 
                [print(i, file=_fw) for i in result] 
            with open('data/sifts.mining.latest.csv', 'w') as _fw: 
                [print(i, file=_fw) for i in result] 
            break

df = pd.read_csv('data/sifts.mining.latest.csv', 
                 names=['pdb_accessionid', 
                        'pdb_chainid',
                        'pdb_resnum', 
                        'pdb_resname', 
                        'unp_accessionid', 
                        'unp_resnum',
                        'unp_resname'],
                 dtype={'pdb_accessionid': str,
                        'pdb_chainid': str,
                        'pdb_resnum': str,
                        'pdb_resname': str,
                        'unp_accessionid': str,
                        'unp_resnum': str,
                        'unp_resname': str}
                )

df.to_sql('sifts', con, if_exists='append',index=False)

pd.read_sql("SELECT name FROM sqlite_master WHERE type ='table';", con)

from IPython.core.display import display, HTML, SVG
import pandas as pd
display(HTML("<style>.container { width:100% !important; }</style>"))  
pd.set_option('display.max_columns', None)
pd.set_option('display.max_colwidth', None) 
df = pd.read_pickle('data/df_latest.pkl') 
df.head()

df['pdb_id_lower']     = df.apply(lambda x: x.pdb_id.lower(), axis=1)  
df['res_num_extented'] = df.apply(lambda r: r.covalent_bond_record.split(',')[6] if r.covalent_bond_record.split(',')[9]=='?' else r.covalent_bond_record.split(',')[6] + r.covalent_bond_record.split(',')[9], axis=1)  
df[['pdb_id_lower','chain_id','res_num_extented']].to_sql('pdb_cov',con,if_exists="replace",index=True)
con.commit()

cur.execute("SELECT * FROM pdb_cov LIMIT 3;")
cur.fetchall()

df_hit = pd.read_sql_query('''SELECT sifts.*       ,         pdb_cov.adduct_id 
                              FROM   sifts       JOIN        pdb_cov 
                              ON     sifts.pdb_accessionid = pdb_cov.pdb_id_lower 
                                 AND sifts.pdb_chainid     = pdb_cov.chain_id 
                                 AND sifts.pdb_resnum      = pdb_cov.res_num_extented''', con)

len(df_hit)

df_hit.to_csv(datetime.datetime.fromtimestamp(time.time()).strftime("data/sifts.matched.csv"),index=False)

cur.execute("DROP TABLE pdb_cov;")
cur.fetchall()

df_hit

df

>sqlite3 c:/sqlite/chinook.db
sqlite> .headers on
sqlite> .mode csv
sqlite> .output data.csv
sqlite> SELECT customerid,
   ...>        firstname,
   ...>        lastname,
   ...>        company
   ...>   FROM customers;
sqlite> .quit

df = pd.read_csv('/scratch/xg590/scratch/sifts.20210826.csv', 
                 names=['pdb_accessionid', 
                        'pdb_chainid',
                        'pdb_resnum', 
                        'pdb_resname', 
                        'unp_accessionid', 
                        'unp_resnum',
                        'unp_resname'],
                 dtype={'pdb_accessionid': str,
                        'pdb_chainid': str,
                        'pdb_resnum': str,
                        'pdb_resname': str,
                        'unp_accessionid': str,
                        'unp_resnum': str,
                        'unp_resname': str}
                ) 
import sqlite3 
con = sqlite3.connect('/scratch/xg590/scratch/sifts.20210826.db')
df.to_sql('sifts', con, if_exists='replace',index=False)
