==> 5.BindingDB.ipynb <==
# https://www.bindingdb.org/bind/chemsearch/marvin/SDFdownload.jsp?all_download=yes
# BindingDB_BindingDB_Inhibition_2021m6.tsv.zip

from rdkit import Chem
from rdkit.Chem import AllChem, PandasTools 
from rdkit.Chem.Draw import IPythonConsole  
import pandas as pd, sqlite3, math
pd.set_option('display.max_columns', None)
#pd.set_option('display.max_rows', None)
pd.set_option('display.max_colwidth', None) 

tsv = '/scratch/xg590/source/BindingDB_All.tsv'
csv = '/scratch/xg590/scratch/BindingDB_All.csv'
with open(tsv, 'r') as fr, open(csv, 'w') as fw:
    fr.readline()
    print('BindingDB Reactant_set_id,Ligand SMILES,UniProt (SwissProt) ID of Target Chain,Ligand InChIKey', file=fw)
    while True:
        line = fr.readline()
        if line:
            line = line.split('\t')
            item = [line[i] for i in [0,1,41,42]] 
            if item[2]:
                mol = Chem.MolFromSmiles(item[1])
                if not mol: continue
                smiles = Chem.MolToSmiles(mol,isomericSmiles=True,canonical=True)
                inchikey = Chem.MolToInchiKey(mol)
                for i in item[2].split(','):
                    print(','.join([item[0], smiles, i, ]), file=fw)
            if item[3]:
                print(line)
                beak
                for i in item[3].split(','):
                    
                    print(','.join([item[0], smiles, i]), file=fw)  
        else:  
            break # executed in 15m 13s

csv = '/scratch/xg590/source/BindingDB_All.csv'
df = pd.read_csv(csv, usecols=['BindingDB Reactant_set_id', 'Ligand InChI Key',
       'UniProt (SwissProt) Primary ID of Target Chain',
       'UniProt (SwissProt) Secondary ID(s) of Target Chain'], low_memory=False)

df = df[df['UniProt (SwissProt) Primary ID of Target Chain'].notna()]
df = df.fillna('N/A')

con = sqlite3.connect('/scratch/xg590/scratch/BindingDB.db') #  :memory:

df[['BindingDB Reactant_set_id', 'Ligand InChI Key',
   'UniProt (SwissProt) Primary ID of Target Chain']].\
   rename(columns={
    'BindingDB Reactant_set_id': 'react_id',
    'Ligand InChI Key':'inchikey',
    'UniProt (SwissProt) Primary ID of Target Chain':'unprotid'
   }).\
   to_sql('bindingdb',con,if_exists="replace",index=False) 

cur = con.cursor()

for _, row in df.iterrows(): 
    react_id = row['BindingDB Reactant_set_id']
    inchikey = row['Ligand InChI Key']
    uniprot  = row['UniProt (SwissProt) Secondary ID(s) of Target Chain'] 
    cur.executemany('''
        insert into bindingdb(react_id, inchikey, unprotid) values (?, ?, ?)
    ''', [(react_id, inchikey, i) for i in uniprot.split(',')])

con.commit()

con.close()

con = sqlite3.connect('/scratch/xg590/scratch/BindingDB.db') #  :memory:
cur = con.cursor()

df = pd.read_pickle('data/df_processed.pkl') 
df[['InChIKey','binder_smiles','unp_accessionid','unp_resnum','record_id']].to_sql('pdbcov',con,if_exists="replace",index=True)  

pd.read_sql('select * from bindingdb limit 1', con)

pd.read_sql('select * from pdbcov limit 1', con)

Mapping = pd.read_sql_query('''SELECT pdbcov.adduct_id, bindingdb.react_id, unp_accessionid, binder_smiles, record_id
                               FROM pdbcov JOIN bindingdb
                               ON     pdbcov.unp_accessionid = bindingdb.unprotid
                                  AND pdbcov.InChIKey        = bindingdb.inchikey''' , con) 
# executed in 39.3s, finished 06:37:50 2021-05-21

Mapping.drop_duplicates(['unp_accessionid','binder_smiles'])

mapping = Mapping.sort_values('record_id')\
                 .drop_duplicates(['record_id'])[['record_id','react_id']]\
                 .set_index('record_id')\
                 .to_dict('dict')['react_id']
df['BindingDB_react_id']=df.apply(lambda r: str(mapping[r.record_id]) if r.record_id in mapping else math.nan,axis=1)
df.to_pickle('data/df_processed.pkl')
